refactor: auto inject all validators
