+++ packages/api/app/Utils/CustomValidators/index.js
@@ -1,0 +1,9 @@
+const cpf = require('./cpf');
+const cnpj = require('./cnpj');
+const exists = require('./exists');
+
+module.exports = {
+	cpf,
+	cnpj,
+	exists,
+};
