+++ packages/api/start/hooks.js
@@ -1,4 +1,3 @@
-/* eslint-disable global-require */
 const { hooks } = require('@adonisjs/ignitor');
 
 const AlgoliaFakeProvider = require('../test/utils/AlgoliaFakeService');
@@ -23,31 +22,11 @@ hooks.before.providersRegistered(() => {
 
 hooks.after.providersBooted(() => {
 	const Validator = use('Validator');
-	const Database = use('Database');
 
-	const existsFn = async (data, field, message, args, get) => {
-		const value = get(data, field);
-		if (!value) {
-			/**
-			 * skip validation if value is not defined. `required` rule
-			 * should take care of it.
-			 */
-			return;
-		}
+	// eslint-disable-next-line global-require
+	const CustomValidators = require('../app/Utils/CustomValidators');
 
-		const [table, column] = args;
-		const row = await Database.table(table)
-			.where(column, value)
-			.first();
-
-		if (!row) {
-			throw message;
-		}
-	};
-
-	const cpf = require('../app/Utils/cpf');
-	const cnpj = require('../app/Utils/cnpj');
-	Validator.extend('exists', existsFn);
-	Validator.extend('cpf', cpf);
-	Validator.extend('cnpj', cnpj);
+	Object.values(CustomValidators).forEach((validator) => {
+		Validator.extend([validator.name], validator);
+	});
 });
